Add
This function has to make a new list and that new list has to have more new lists inside it. And we're going to need to loop over our old lists of lists while doing that.

One of the trickier parts of this problem is the fact that you'll need to loop over two lists at the same time.